Allow telemetry to be initialised on Serial0 only and allow Serial3 to optionally not be configured at all , or to be used for something entirely different.
On APM2 hardware, the usb multiplexer, and existing code based abound USB_MUX_PIN does all the heavy lfting, and this "just works".  :-)

On APM1 hardware, If we connect the telemetry (Xbees) to Serial0 instead of Serial3, we can free up Serial3 for other things.
The Serial3 is no longer tied up doing telemetry if these defines are set:

Don't forget to manually change the electrical connection/s of your Xbee to connect to the FTDI pins on the "red board" if you use this setup.

Also, this means that Serial3 can be used for something else ( like a second GPS module, or a redundant telemetry link using a different system? )